[UI 수정] 맥락 인증, 금액 별 인증 텍스트에 폰트 적용

+++ app/src/main/java/finotek/global/dev/talkbank_ca/base/mvp/event/AccuracyMeasureEvent.java
@@ -11,7 +11,7 @@ public class AccuracyMeasureEvent implements IEvent {
 		this.accuracy = accuracy;
 	}
 
-	public int getAccuracy() {
-		return (int) accuracy;
+	public double getAccuracy() {
+		return accuracy;
 	}
 }
